+++ tests/common.rs
@@ -35,9 +35,9 @@ pub fn expect_assert(expr: bool) -> Result<(), String> {
     }
 }
 
-pub fn expect_assert_eq<T: PartialEq>(expr1: T, expr2: T) -> Result<(), String>
+pub fn expect_assert_eq<T>(expr1: T, expr2: T) -> Result<(), String>
 where
-    T: Debug,
+    T: Debug + PartialEq,
 {
     if expr1 != expr2 {
         Err(format!(